requirements.txt
flask==2.3.3
pandas==2.2.3
numpy<2
pyarrow==12.0.0
psycopg2-binary==2.9.7